pulled out rax_auth and made it configurable
